+++ 04 Implementation/src/main/java/view/java_fx/OpretForloeb.fxml
@@ -12,11 +12,11 @@
       <ImageView id="headerimage" fitHeight="42.0" fitWidth="600.0" pickOnBounds="true" preserveRatio="true" />
       <VBox id="menuvbox" layoutY="40.0" prefHeight="361.0" prefWidth="120.0">
          <children>
-            <Button id="forsidebtn" mnemonicParsing="false" prefHeight="26.0" prefWidth="137.0" text="Forside" />
+            <Button id="forsidebtn" onAction="#forside" mnemonicParsing="false" prefHeight="26.0" prefWidth="137.0" text="Forside" />
             <Button id="opretbrugerbtn" layoutX="10.0" layoutY="10.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="133.0" text="Opret bruger" />
             <Button id="logindbtn" layoutX="10.0" layoutY="36.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="149.0" text="Log ind" />
             <Button id="opretforloebbtn" layoutX="10.0" layoutY="62.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="134.0" text="Opret forløb" />
-            <Button id="mineforloebbtn" layoutX="10.0" layoutY="88.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="162.0" text="Mine forløb" />
+            <Button id="mineforloebbtn" onAction="#mineForloebSide" layoutX="10.0" layoutY="88.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="162.0" text="Mine forløb" />
             <Button id="testbtn" layoutX="10.0" layoutY="114.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="140.0" text="Test" />
             <Button id="traeningsprogrambtn" layoutX="10.0" layoutY="140.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="154.0" text="Træningsprogram" />
          </children>
