+++ 04 Implementation/src/main/java/view/java_fx/OpretForloebController.java
@@ -7,11 +7,13 @@ import javafx.scene.control.ComboBox;
 import javafx.stage.Stage;
 import model.datastrukture.Behandler;
 import model.datastrukture.Patient;
+import model.persistence.firebase.FirebaseDAO;
 
 import java.io.IOException;
 import java.util.ArrayList;
 
-public class OpretForloebController {
+public class OpretForloebController extends MenuController {
+    private FirebaseDAO firebaseDAO;
     private Main main;
     private Stage stage;
     //fxml elements
@@ -21,6 +23,9 @@ public class OpretForloebController {
     private ArrayList<Patient> patienter;
 
     public void load(Main main, Stage stage){
+        if (firebaseDAO==null){
+            firebaseDAO = new FirebaseDAO();
+        }
         //setting references
         this.stage=stage;
         this.main = main;
@@ -34,7 +39,8 @@ public class OpretForloebController {
     //TODO VIEW MODEL SKAL BRUGES NU :!:!:!!::!!
     private void setBehandlere() {
         try {
-            behandlere = main.getFirebaseDAO().hentBehandlere();
+            //behandlere = main.getFirebaseDAO().hentBehandlere();
+            behandlere = firebaseDAO.hentBehandlere();
             for (Behandler behandler : behandlere){
                 System.out.println("BEHANDLER LOADED: "+behandler.getEmail());
                 behandlercombobox.getItems().add(behandler.getNavn());
@@ -47,7 +53,8 @@ public class OpretForloebController {
 
     private void setPatienter(){
         try {
-            patienter = main.getFirebaseDAO().hentPatienter(false);
+            //patienter = main.getFirebaseDAO().hentPatienter(false);
+            patienter = firebaseDAO.hentPatienter(false);
             for (Patient patient : patienter){
                 System.out.println("PATIENT LOADED: "+patient.getEmail());
                 patientcombobox.getItems().add(patient.getNavn());
@@ -61,7 +68,7 @@ public class OpretForloebController {
     public void opretForloeb(ActionEvent event) throws IOException {
         System.out.println("Behandler: "+behandlere.get(behandlercombobox.getSelectionModel().getSelectedIndex()).getNavn()+" Patient: "+patienter.get(patientcombobox.getSelectionModel().getSelectedIndex()).getNavn());
         try {
-            main.getFirebaseDAO().opretForloeb(behandlere.get(behandlercombobox.getSelectionModel().getSelectedIndex()), patienter.get(patientcombobox.getSelectionModel().getSelectedIndex()));
+            firebaseDAO.opretForloeb(behandlere.get(behandlercombobox.getSelectionModel().getSelectedIndex()), patienter.get(patientcombobox.getSelectionModel().getSelectedIndex()));
             System.out.println("successfuldt oprettet forløb!");
             //back to front page
             main.start(stage);
